+++ docs/src/markdown/about/changelog.md
@@ -10,6 +10,7 @@
 - **NEW**: `xyz` now refers to XYZ D65. CSS `#!css-color color()` function now specifies D65 color as either
   `#!css-color color(xyz x y z)` or `#!css-color color(xyz-d65 x y z)`. XYZ D50 is now specified as
   `#!css-color color(xyz-D50 x y z)`.
+- **NEW**: Add CIELUV and CIELCH~uv~ D65 variants.
 
 ## 0.2.0